Moved date filter into a region object.
