+++ test_date_filter.py
@@ -65,13 +65,13 @@ class TestSearchDates:
         feedback_pg = feedback_page.FeedbackPage(testsetup)
 
         feedback_pg.go_to_feedback_page()
-        Assert.equal(feedback_pg.get_current_days(), u"\u221e")
+        Assert.equal(feedback_pg.date_filter.current_days, u"\u221e")
 
         day_filters = ((1, "1d", "Last day"), (7, "7d", "Last 7 days"), (30, "30d", "Last 30 days"))
         for days in day_filters:
-            Assert.equal(feedback_pg.get_days_tooltip(days[1]), days[2])
-            feedback_pg.click_days(days[1])
-            Assert.equal(feedback_pg.get_current_days(), days[1])
+            Assert.equal(feedback_pg.date_filter.get_days_tooltip(days[1]), days[2])
+            feedback_pg.date_filter.click_days(days[1])
+            Assert.equal(feedback_pg.date_filter.current_days, days[1])
             start_date = date.today() - timedelta(days=days[0])
             # The format for a date when using preset filters is different to using the custom search. See bug 616306 for details.
             Assert.equal(feedback_pg.date_start_from_url, start_date.strftime('%Y-%m-%d'))
@@ -88,12 +88,12 @@ class TestSearchDates:
         feedback_pg = feedback_page.FeedbackPage(testsetup)
 
         feedback_pg.go_to_feedback_page()
-        Assert.equal(feedback_pg.get_custom_dates_tooltip(), "Custom")
+        Assert.equal(feedback_pg.date_filter.custom_dates_tooltip, "Custom")
 
         start_date = date.today() - timedelta(days=3)
         end_date = date.today() - timedelta(days=1)
 
-        feedback_pg.filter_by_custom_dates_using_datepicker(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_datepicker(start_date, end_date)
         # The format for a date when using preset filters is different to using the custom search. See bug 616306 for details.
         Assert.equal(feedback_pg.date_start_from_url, start_date.strftime('%m%%2F%d%%2F%Y'))
         Assert.equal(feedback_pg.date_end_from_url, end_date.strftime('%m%%2F%d%%2F%Y'))
@@ -103,12 +103,12 @@ class TestSearchDates:
         day_filters = ((1, "1d"), (7, "7d"), (30, "30d"))
         for days in day_filters:
             start_date = date.today() - timedelta(days=days[0])
-            feedback_pg.filter_by_custom_dates_using_datepicker(start_date, date.today())
-            Assert.false(feedback_pg.is_custom_date_filter_visible())
+            feedback_pg.date_filter.filter_by_custom_dates_using_datepicker(start_date, date.today())
+            Assert.false(feedback_pg.date_filter.is_custom_date_filter_visible)
             # The format for a date when using preset filters is different to using the custom search. See bug 616306 for details.
             Assert.equal(feedback_pg.date_start_from_url, start_date.strftime('%m%%2F%d%%2F%Y'))
             Assert.equal(feedback_pg.date_end_from_url, date.today().strftime('%m%%2F%d%%2F%Y'))
-            Assert.equal(feedback_pg.get_current_days(), days[1])
+            Assert.equal(feedback_pg.date_filter.current_days, days[1])
 
     def test_feedback_custom_date_filter_with_random_alphabet(self, testsetup):
         """
@@ -125,13 +125,13 @@ class TestSearchDates:
         start_date = ''.join(random.sample(letters, 8))
         end_date = ''.join(random.sample(letters, 8))
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, '')
         Assert.equal(feedback_pg.date_end_from_url, '')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, '')
-        Assert.equal(feedback_pg.custom_end_date, '')
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, '')
+        Assert.equal(feedback_pg.date_filter.custom_end_date, '')
 
     def test_feedback_custom_date_filter_with_random_numbers(self, testsetup):
         """
@@ -147,17 +147,17 @@ class TestSearchDates:
         start_date = random.randint(10000000, 50000000)
         end_date = random.randint(50000001, 99999999)
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, str(start_date))
         Assert.equal(feedback_pg.date_end_from_url, str(end_date))
 
         Assert.equal(feedback_pg.warning_heading, 'No search results found.')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, str(start_date))
-        Assert.equal(feedback_pg.custom_end_date, str(end_date))
-        Assert.equal(feedback_pg.custom_date_first_error, 'Enter a valid date.')
-        Assert.equal(feedback_pg.custom_date_second_error, 'Enter a valid date.')
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, str(start_date))
+        Assert.equal(feedback_pg.date_filter.custom_end_date, str(end_date))
+        Assert.equal(feedback_pg.date_filter.custom_date_first_error, 'Enter a valid date.')
+        Assert.equal(feedback_pg.date_filter.custom_date_second_error, 'Enter a valid date.')
 
     def test_feedback_custom_date_filter_with_invalid_dates(self, testsetup):
         """
@@ -173,17 +173,17 @@ class TestSearchDates:
         start_date = "00/00/0000"
         end_date = "00/00/0000"
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, string.replace(start_date, '/', '%2F'))
         Assert.equal(feedback_pg.date_end_from_url, string.replace(end_date, '/', '%2F'))
 
         Assert.equal(feedback_pg.warning_heading, 'No search results found.')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, start_date)
-        Assert.equal(feedback_pg.custom_end_date, end_date)
-        Assert.equal(feedback_pg.custom_date_first_error, 'Enter a valid date.')
-        Assert.equal(feedback_pg.custom_date_second_error, 'Enter a valid date.')
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, start_date)
+        Assert.equal(feedback_pg.date_filter.custom_end_date, end_date)
+        Assert.equal(feedback_pg.date_filter.custom_date_first_error, 'Enter a valid date.')
+        Assert.equal(feedback_pg.date_filter.custom_date_second_error, 'Enter a valid date.')
 
     def test_feedback_custom_date_filter_with_future_dates(self, testsetup):
         """
@@ -200,15 +200,15 @@ class TestSearchDates:
         start_date = "01/01/2021"
         end_date = "01/01/2031"
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, string.replace(start_date, '/', '%2F'))
         Assert.equal(feedback_pg.date_end_from_url, string.replace(end_date, '/', '%2F'))
 
         Assert.equal(feedback_pg.warning_heading, 'No search results found.')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, start_date)
-        Assert.equal(feedback_pg.custom_end_date, end_date)
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, start_date)
+        Assert.equal(feedback_pg.date_filter.custom_end_date, end_date)
 
     @xfail(reason="Bug 645850 - [input-stage] Internal Server Error - OverflowError: mktime argument out of range")
     def test_feedback_custom_date_filter_with_future_start_date(self, testsetup):
@@ -226,15 +226,15 @@ class TestSearchDates:
         start_date = "01/01/2900"
         end_date = ""
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, string.replace(start_date, '/', '%2F'))
         Assert.equal(feedback_pg.date_end_from_url, string.replace(end_date, '/', '%2F'))
 
         Assert.equal(feedback_pg.warning_heading, 'No search results found.')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, start_date)
-        Assert.equal(feedback_pg.custom_end_date, end_date)
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, start_date)
+        Assert.equal(feedback_pg.date_filter.custom_end_date, end_date)
 
     @xfail(reason="Bug 645850 - [input-stage] Internal Server Error - OverflowError: mktime argument out of range")
     def test_feedback_custom_date_filter_with_future_end_date(self, testsetup):
@@ -252,15 +252,15 @@ class TestSearchDates:
         start_date = ""
         end_date = "01/01/2900"
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, string.replace(start_date, '/', '%2F'))
         Assert.equal(feedback_pg.date_end_from_url, string.replace(end_date, '/', '%2F'))
 
         Assert.true(feedback_pg.is_text_present('Search Results'))
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, start_date)
-        Assert.equal(feedback_pg.custom_end_date, end_date)
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, start_date)
+        Assert.equal(feedback_pg.date_filter.custom_end_date, end_date)
 
     def test_feedback_custom_date_filter_with_end_date_lower_than_start_date(self, testsetup):
         """
@@ -276,15 +276,15 @@ class TestSearchDates:
         start_date = date.today() - timedelta(days=1)
         end_date = date.today() - timedelta(days=3)
 
-        feedback_pg.filter_by_custom_dates_using_datepicker(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_datepicker(start_date, end_date)
         # The format for a date when using preset filters is different to using the custom search. See bug 616306 for details.
         Assert.equal(feedback_pg.date_start_from_url, start_date.strftime('%m%%2F%d%%2F%Y'))
         Assert.equal(feedback_pg.date_end_from_url, end_date.strftime('%m%%2F%d%%2F%Y'))
         # TODO: Check results are within the expected date range, possibly by navigating to the first/last pages and checking the final result is within range. Currently blocked by bug 615844.
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, start_date.strftime('%m/%d/%Y'))
-        Assert.equal(feedback_pg.custom_end_date, end_date.strftime('%m/%d/%Y'))
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, start_date.strftime('%m/%d/%Y'))
+        Assert.equal(feedback_pg.date_filter.custom_end_date, end_date.strftime('%m/%d/%Y'))
 
     def test_feedback_custom_date_filter_with_ydm_format(self, testsetup):
         """
@@ -300,13 +300,13 @@ class TestSearchDates:
         start_date = '2011-22-04'
         end_date = ''
 
-        feedback_pg.filter_by_custom_dates_using_keyboard(start_date, end_date)
+        feedback_pg.date_filter.filter_by_custom_dates_using_keyboard(start_date, end_date)
         Assert.equal(feedback_pg.date_start_from_url, string.replace(start_date, '-', ''))
         Assert.equal(feedback_pg.date_end_from_url, '')
 
         Assert.equal(feedback_pg.warning_heading, 'No search results found.')
 
-        feedback_pg.click_custom_dates()
-        Assert.equal(feedback_pg.custom_start_date, string.replace(start_date, '-', ''))
-        Assert.equal(feedback_pg.custom_end_date, '')
-        Assert.equal(feedback_pg.custom_date_only_error, 'Enter a valid date.')
+        feedback_pg.date_filter.click_custom_dates()
+        Assert.equal(feedback_pg.date_filter.custom_start_date, string.replace(start_date, '-', ''))
+        Assert.equal(feedback_pg.date_filter.custom_end_date, '')
+        Assert.equal(feedback_pg.date_filter.custom_date_only_error, 'Enter a valid date.')
